construction/project-master/requirements.txt
django
django-crispy-forms
stripe
django-allauth
pillow
